+++ esp32/lib/dirtviz/src/dirtviz.cpp
@@ -8,6 +8,8 @@
 #include "dirtviz.hpp"
 #include <Arduino.h>
 
+#include <cstdio>
+
 Dirtviz::Dirtviz(const char *url, const uint16_t &port)
     : url(nullptr), response(nullptr) {
   // set parameters