+++ proto/python/src/soil_power_sensor_protobuf/proto/encode.py
@@ -10,7 +10,13 @@ Each type of measurement has a corresponding encoding function as follows:
     Teros12Measurement -> encode_teros12_measurement()
 """
 
-from .soil_power_sensor_pb2 import Measurement, Response, UserConfiguration, EnabledSensor, Uploadmethod
+from .soil_power_sensor_pb2 import (
+    Measurement,
+    Response,
+    UserConfiguration,
+    EnabledSensor,
+    Uploadmethod,
+)
 
 
 def encode_response(success: bool = True) -> bytes:
@@ -133,8 +139,61 @@ def encode_phytos31_measurement(
     # phytos31
     meas.phytos31.voltage = voltage
     meas.phytos31.leaf_wetness = leaf_wetness
+    return meas.SerializeToString()
 
     return meas.SerializeToString()
+
+
+def encode_bme280_measurement(
+    ts: int,
+    cell_id: int,
+    logger_id: int,
+    pressure: int,
+    temperature: int,
+    humidity: int,
+):
+    """Encodes a BME280Measurement within the Measurement message
+
+    The following raw values correspond to the following SI units
+
+    *Raw*
+
+    pressure: 98473
+    temperature: 2275
+    humidity: 43600
+
+    *SI Units*
+
+    pressure: 9847.3 hPa
+    temperature: 22.75 C
+    humidity: 43.600 %
+
+    Args:
+        ts: Timestamp in unix epochs
+        cell_id: Cell Id from Dirtviz
+        logger_id: Logger Id from Dirtviz
+        pressure: Ambient pressure
+        temperature: Ambient temperature
+        humidity: Relative humidity
+
+    Returns:
+        Serialized BME280 Measurement
+    """
+
+    meas = Measurement()
+
+    # metadata
+    meas.meta.ts = ts
+    meas.meta.cell_id = cell_id
+    meas.meta.logger_id = logger_id
+
+    # bme280
+    meas.bme280.pressure = pressure
+    meas.bme280.temperature = temperature
+    meas.bme280.humidity = humidity
+
+    return meas.SerializeToString()
+
 
 def encode_user_configuration(
     logger_id: int,
@@ -149,11 +208,8 @@ def encode_user_configuration(
     WiFi_SSID: str,
     WiFi_Password: str,
     API_Endpoint_URL: str,
-    API_Endpoint_Port: int
+    API_Endpoint_Port: int,
 ) -> bytes:
-
-
-
     """Encodes a UserConfiguration message
 
     Args:
@@ -188,14 +244,14 @@ def encode_user_configuration(
         user_config.Upload_method = Uploadmethod.WiFi
     else:
         raise ValueError("Invalid Upload_method: must be 'LoRa' or 'WiFi'")
-    
+
     # Check for duplicates in Enabled_sensors
     seen_sensors = set()
     for sensor in Enabled_sensors:
         if sensor in seen_sensors:
             raise ValueError(f"Duplicate sensor found: {sensor}")
         seen_sensors.add(sensor)
-        
+
         # Append to enabled_sensors based on the enum mapping
         if sensor.lower() == "voltage":
             user_config.enabled_sensors.append(EnabledSensor.Voltage)
@@ -212,7 +268,7 @@ def encode_user_configuration(
     user_config.Voltage_Slope = Voltage_Slope
     user_config.Voltage_Offset = Voltage_Offset
     user_config.Current_Slope = Current_Slope
-    user_config.Current_Offset = Current_Offset    
+    user_config.Current_Offset = Current_Offset
     user_config.WiFi_SSID = WiFi_SSID
     user_config.WiFi_Password = WiFi_Password
     user_config.API_Endpoint_URL = API_Endpoint_URL