+++ stm32/Src/main.c
@@ -37,6 +37,7 @@
 #include "ads.h"
 #include "sdi12.h"
 #include "phytos31.h"
+#include "bme280_sensor.h"
 #include "rtc.h"
 #include "sensors.h"
 /* USER CODE END Includes */
@@ -120,6 +121,10 @@ int main(void)
   SensorsAdd(ADC_measure);
   //SensorsAdd(SDI12_Teros12Measure);
   //SensorsAdd(Phytos31_measure);
+  
+  //BME280Init();
+  //SensorsAdd(BME280Measure); 
+
 
 
   /* USER CODE END 2 */
@@ -220,7 +225,8 @@ void assert_failed(uint8_t *file, uint32_t line)
 {
   /* USER CODE BEGIN 6 */
   /* User can add his own implementation to report the file name and line number,
-     ex: printf("Wrong parameters value: file %s on line %d\r\n", file, line) */
+   * example: printf("Wrong parameters value: file %s on line %d\r\n", file, line)
+   */
   /* USER CODE END 6 */
 }
 #endif /* USE_FULL_ASSERT */