+++ CHANGELOG.md
@@ -12,6 +12,20 @@ firmware versions for a hardware version.
 
 ### Added
 
+- Support for bme280 sensor [#87](i87), [#99](i99)
+- Support for MB85RC1MT fram chip [#136](i136)
+
+### Changed
+
+- Measurement mode of bme280 to forced [#99](i99)
+
+[i87]: https://github.com/jlab-sensing/ENTS-node-firmware/issues/87
+[i99]: https://github.com/jlab-sensing/ENTS-node-firmware/issues/99
+[i136]: https://github.com/jlab-sensing/ENTS-node-firmware/issues/136
+
+## [2.2.0] - 2024-12-13
+
+### Added
 - Contributing resources ([Contributing](./CONTRIBUTING.md),
   [Code ofConduct](./CODE_OF_CONDUCT.md), and Changelog)
 - Issue and PR templates
@@ -19,6 +33,7 @@ firmware versions for a hardware version.
 - Transmit buffer [#41](i41)
 - RTC time synchronization support through LoRaWAN [#48](i48)
 - Calibration support [#50](i50)
+- Minimum working example for configuration GUI
 
 ### Changed